+++ components/main/Main.style.tsx
