+++ components/main/Main.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import styled from 'styled-components';
 
 // Styled
-import { StyledMain } from '../styled/components/StyledMain';
+import { StyledMain } from './Main.style';
 
 function Main(props) {
   return <StyledMain {...props} />;
