+++ components/main/MainList.style.tsx
