+++ components/main/MainSection.tsx
@@ -1,0 +1,8 @@
+import * as React from 'react';
+import { StyledMovieSection } from './MainSection.style';
+
+function MovieSection(props) {
+  return <StyledMovieSection {...props} />;
+}
+
+export default MovieSection;
