+++ components/main/MainSectionBg.tsx
@@ -1,0 +1,8 @@
+import * as React from 'react';
+import { StyledMovieSectionBg } from './MainSection.style';
+
+function MovieSectionBg({ bgUrl, $sectionBg }) {
+  return <StyledMovieSectionBg bgUrl={bgUrl} ref={$sectionBg} />;
+}
+
+export default MovieSectionBg;
