+++ components/styled/global/StyledReset.tsx
