+++ components/styled/global/StyledVariable.ts
